MF: Don't remove a non-empty folder

+++ ManagedFolders/ManagedFolders.cpp
@@ -152,8 +152,10 @@ private:
                     }
                 }
 
-                if (common_folders.empty())
-                    fs::remove(sub_path2);
+                if (common_folders.empty()) {
+                    if (fs::is_empty(sub_path2))
+                        fs::remove(sub_path2);
+                }
                 else
                     generate_all_folders_synch_script(common_folders, sub_path2, location_path1, location_path2);
             }
